Change speed from increasing every round to fixed, and the number of illuminated targets will go up each wave.
As we (event organizers) work on the contest we've run into issues getting targets to go fast on a fixed track without flying off the rail. For this reason we're switching the contest to use fixed-speed targets and instead require more hits per target per wave. These rules are still in progress and community input is welcomed.